+++ xbmc/StringUtils.h
@@ -35,6 +35,7 @@
 #include "DateTime.h"
 #include "StdString.h"
 #include <vector>
+#include <stdint.h>
 
 class StringUtils
 {
@@ -42,12 +43,12 @@ public:
   static void JoinString(const CStdStringArray &strings, const CStdString& delimiter, CStdString& result);
   static int SplitString(const CStdString& input, const CStdString& delimiter, CStdStringArray &results, unsigned int iMaxStrings = 0);
   static int FindNumber(const CStdString& strInput, const CStdString &strFind);
-  static long long AlphaNumericCompare(const char *left, const char *right);
+  static int64_t AlphaNumericCompare(const char *left, const char *right);
   static long TimeStringToSeconds(const CStdString &timeString);
   static void RemoveCRLF(CStdString& strLine);
   static void SecondsToTimeString( long lSeconds, CStdString& strHMS, TIME_FORMAT format = TIME_FORMAT_GUESS);
   static bool IsNaturalNumber(const CStdString& str);
-  static CStdString SizeToString(__int64 size);
+  static CStdString SizeToString(int64_t size);
   static const CStdString EmptyString;
   static size_t FindWords(const char *str, const char *wordLowerCase);
   static int FindEndBracket(const CStdString &str, char opener, char closer, int startPos = 0);
